move test_tankoubon_edit_page_save, remove xfail flag
